Refactor ContentArea component to utilize updated Content and Notes types; implement removeContent function in useContent hook for content deletion with success/error notifications.
