+++ types/content-type.ts
@@ -37,3 +37,41 @@ export type ShareData = {
   contents: SharedContent[];
   notes: SharedNote[];
 };
+
+interface ContentTag {
+  contentId: string;
+  tagsId: string;
+  tags: {
+    id: string;
+    title: string;
+  };
+}
+
+interface NotesTag {
+  notesId: string;
+  tagsId: string;
+  tags: {
+    id: string;
+    title: string;
+  };
+}
+
+export interface Content {
+  id: string;
+  title: string;
+  type: ContentType;
+  summary: string;
+  link: string;
+  createdAt: string;
+  ContentTags: ContentTag[];
+  spacesId?: string | null;
+}
+
+export interface Notes {
+  id: string;
+  title: string;
+  description: string;
+  createdAt: string;
+  NotesTags: NotesTag[];
+  spacesId?: string | null;
+}
